pyproject.toml
[project]
name = "figma-cc"
version = "0.1.0"
description = "Figma plugin + FastAPI backend for AI-powered design feedback from synthetic personas"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.41",
    "fastapi>=0.133.0",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
